requirements.txt
scrapy==2.11.2
scrapy-playwright==0.0.33
playwright==1.43.0
beautifulsoup4
lxml
readability-lxml
python-dotenv
